closed https://github.com/BetaMasaheft/Documentation/issues/985
also for languages a static list languages.xml has been created from the data.
This will not be updated when new languages are registered but the utility script is stored in collection utilities for the purpose of updating this file.
app.xql and apprest.xql have been updated to use this