support for deleted files
https://github.com/BetaMasaheft/Documentation/issues/1401
deleted.html is now a basic list of deleted files with permalinks, based on the existing deleted.xml. this has to receive more information, namely source and date of change for the permalinks to be retrievable. also the behavious or permanentItems.xqm has been improved to show also these records which do not exist any more if requested with a permanent identifier. permanentID.js has been updated to be used on deleted.html as well, with a generalization of the id structure instead of a straigth forward id